+++ exporters/otlp/otlplog/otlploggrpc/config_test.go
@@ -334,6 +334,28 @@ func TestNewConfig(t *testing.T) {
 			},
 		},
 		{
+			name: "OptionEndpointURLWithoutScheme",
+			options: []Option{
+				WithEndpointURL("//env.endpoint:8080/prefix"),
+			},
+			want: config{
+				endpoint: newSetting("env.endpoint:8080"),
+				retryCfg: newSetting(defaultRetryCfg),
+				timeout:  newSetting(defaultTimeout),
+			},
+		},
+		{
+			name: "EnvEndpointWithoutScheme",
+			envars: map[string]string{
+				"OTEL_EXPORTER_OTLP_LOGS_ENDPOINT": "//env.endpoint:8080/prefix",
+			},
+			want: config{
+				endpoint: newSetting("env.endpoint:8080"),
+				retryCfg: newSetting(defaultRetryCfg),
+				timeout:  newSetting(defaultTimeout),
+			},
+		},
+		{
 			name: "DefaultEndpointWithEnvInsecure",
 			envars: map[string]string{
 				"OTEL_EXPORTER_OTLP_LOGS_INSECURE": "true",
@@ -362,6 +384,21 @@ func TestNewConfig(t *testing.T) {
 			name: "OptionEndpointURLWithoutSchemeWithEnvInsecure",
 			options: []Option{
 				WithEndpointURL("//env.endpoint:8080/prefix"),
+			},
+			envars: map[string]string{
+				"OTEL_EXPORTER_OTLP_LOGS_INSECURE": "true",
+			},
+			want: config{
+				endpoint: newSetting("env.endpoint:8080"),
+				insecure: newSetting(true),
+				retryCfg: newSetting(defaultRetryCfg),
+				timeout:  newSetting(defaultTimeout),
+			},
+		},
+		{
+			name: "OptionEndpointWithEnvInsecure",
+			options: []Option{
+				WithEndpoint("env.endpoint:8080"),
 			},
 			envars: map[string]string{
 				"OTEL_EXPORTER_OTLP_LOGS_INSECURE": "true",